rebase wallet-core from trustwallet (#87)
* Add Cronos to registry (#2040)

* Add Cronos to registry
* cleanup

* Add TWPBKDF2 C interface (#2042)

* Add TWPBKDF2 C interface
* Fix tests

* Add Smart Bitcoin Cash (#1790)

* Add Smart Bitcoin Cash

* Update TWAnyAddressTests.cpp

add smartBCH testcase

* Change TWTransactionCompiler to struct (#2073)

TWTransactionCompiler only has static methods, and it is not an
object, so change ti to struct, and then wallet-core-python can
parse wallet-core header correctly.

* [Polkadot] Allow network value > 64 for SS58Address (#2065)

* Support full Substrate address (network >= 64)
* Move source, add test
* Move update
* Use constants
* Unify network type (uint32)
* Change bit operations
* Move implementation to .cpp file
* Adjust bit manipulation
* iOS compile fix (numeric type casts)
* Test adjustments

* Fiix MacOS compile error; pointer-to-numerical conversion (#2083)

* Expose public key for account (#2068)

* Expose public key for account
* More typo fixes
* cleanup

* Stellar : Support Claimable Balance (#2057)

* Stellar proto messages for Claimable Balance

* Claimant (in proto)

* Support Claimable Balance messages

* Correct encoding, real CreateClaimable test tx

* Add real Claim test tx

* Final touchup; proto comment, coverage metric

* Minor renames, error handling, review comments

* Comment cleanup

* [Thorchain] Support protobuf serialization messages (Stargate) (#2092)

* Thorchain: Support protobuf serialization messages (Stargate)

* Update Kotlin test

* Update thorchain swift test (#2095)

* Exit early if release id is null (#2096)

* GetExtendedKey with custom account (#2085)

* GetExtendedKey with custom account

* Don't use default param value in C impl (review comment)

* Add KCC Chain (#2098)

* Add KCC Chain
* [fix] rename KCCChain
* [fix]  rename kcc chain name
* [fix] android test fail
* [fix] android TestAddress Failure

* clean up not used xxhash64 (#2109)

* Fix Cronos address data (#2110)

* TWAccount creation fixed (#2111)

* Cardano (Shelley) support (#1944)

* Updated Shelley address handling prototype

* Cleanup, PrivateKey, HDWallet

* AddressV3: cleanup, update, tests

* Update PrivateKey tests

* Test data update

* Add extra address tests

* Simple signing test

* Proto file, basic frame for Signing

* Proto cleanup

* Implement transaction planning

* iOS compile fix, numerical type

* Android address derivation test update

* Add plan in SigningInput, txid in SigningOutput

* Android address test update

* iOS compile fix, numerical types

* iOS test result fix

* Cbor: Support for null

* Update explorer to cardanoscan.io

* Rework: change address and multiple private keys

* Fix staking key generation in Cardano key derivation

* Signature handling

* Add real-world transfer example

* Update iOS address test

* Swift signing test

* Swift test fix

* Swift fix 2

* Implement AnySign

* Add Kotlin signing test

* Kotlin test fix, imports

* AnySigner test

* Test fixes, coverage

* Coverage, Kotlin test fix

* Kotlin quick fix

* Kt fix 4

* Kt fix 5

* Minor cleanup, coverage

* Finalize amount type; cleanup

* Max amount support

* Minor, review comments

* Test cleanup

* Add test with legacy address

* Add missing initializations to AddressV3

* Sending from/to legacy tests

* Case for inputs with tokens, treat as error

* AnyPlan support

* fix some conversion warning

* fix cosmos type issue

* fix ss58address refer issue

* fix proto conflicts

+++ tests/TransactionCompilerTests.cpp
@@ -1609,8 +1609,8 @@ TEST(TransactionCompiler, OntCompileWithSignatures) {
         coin, ongTxInputData, {ongSignature}, {publicKeyData});
     auto ontOutput = Ontology::Proto::SigningOutput();
     auto ongOutput = Ontology::Proto::SigningOutput();
-    ontOutput.ParseFromArray(ontOutputData.data(), ontOutputData.size());
-    ongOutput.ParseFromArray(ongOutputData.data(), ongOutputData.size());
+    ontOutput.ParseFromArray(ontOutputData.data(), (int)ontOutputData.size());
+    ongOutput.ParseFromArray(ongOutputData.data(), (int)ongOutputData.size());
     const auto ontExpectedTx =
         "00d101000000ac0d000000000000204e000000000000ca18ec37ac94f19588926a5302ded54cd909a76e7100c6"
         "6b14ca18ec37ac94f19588926a5302ded54cd909a76e6a7cc8149e21dda3257e18eb033d9451dda1d9ac8bcfa4"
